refactor(kernel): simplify IDT and ISR headers
