--- src/kernel/arch/x86/interrupts.h
@@ -1,4 +1,0 @@
-#pragma once
-
-void
-interrupts_init();
